Move SchemaConfig to common processors package
