+++ dagger-core/src/main/java/io/odpf/dagger/core/processors/common/FetchOutputDecorator.java
@@ -3,7 +3,6 @@ package io.odpf.dagger.core.processors.common;
 import com.google.protobuf.Descriptors;
 import io.odpf.dagger.common.core.StencilClientOrchestrator;
 import io.odpf.dagger.common.serde.proto.protohandler.ProtoHandlerFactory;
-import io.odpf.dagger.core.processors.external.SchemaConfig;
 import io.odpf.dagger.core.processors.types.MapDecorator;
 import org.apache.flink.api.common.typeinfo.TypeInformation;
 import org.apache.flink.api.common.typeinfo.Types;
