+++ dagger-core/src/main/java/io/odpf/dagger/core/processors/internal/processor/InternalConfigHandlerFactory.java
@@ -1,7 +1,7 @@
 package io.odpf.dagger.core.processors.internal.processor;
 
 import io.odpf.dagger.core.processors.ColumnNameManager;
-import io.odpf.dagger.core.processors.external.SchemaConfig;
+import io.odpf.dagger.core.processors.common.SchemaConfig;
 import io.odpf.dagger.core.processors.internal.InternalSourceConfig;
 import io.odpf.dagger.core.processors.internal.processor.sql.SqlConfigTypePathParser;
 import io.odpf.dagger.core.processors.internal.processor.constant.ConstantInternalConfigProcessor;
