+++ dagger-core/src/main/java/io/odpf/dagger/core/processors/internal/processor/function/FunctionInternalConfigProcessor.java
@@ -2,7 +2,7 @@ package io.odpf.dagger.core.processors.internal.processor.function;
 
 import io.odpf.dagger.core.processors.ColumnNameManager;
 import io.odpf.dagger.core.processors.common.RowManager;
-import io.odpf.dagger.core.processors.external.SchemaConfig;
+import io.odpf.dagger.core.processors.common.SchemaConfig;
 import io.odpf.dagger.core.processors.internal.InternalSourceConfig;
 import io.odpf.dagger.core.processors.internal.processor.InternalConfigProcessor;
