+++ dagger-core/src/test/java/io/odpf/dagger/core/processors/external/ExternalPostProcessorTest.java
@@ -1,5 +1,6 @@
 package io.odpf.dagger.core.processors.external;
 
+import io.odpf.dagger.core.processors.common.SchemaConfig;
 import org.apache.flink.streaming.api.datastream.DataStream;
 
 import io.odpf.stencil.client.StencilClient;
